output_pin: Move template evaluation code from led.py to output_pin.py
